product done

+++ src/main/java/store/product/Product.java
@@ -1,6 +1,6 @@
 package store.product;
 
-import java.util.Date;
+
 
 import lombok.Builder;
 import lombok.Data;
@@ -13,6 +13,6 @@ public class Product {
 
     private String id;
     private String name;
-    private Double price;
+    private Float price;
     private String unit;
 }
